Create database tables if they dont exist
Index.php now checks if the 'Movies' table exists in the database. If it does not exist, then it is assumed that either this is a new installation and the tables have not been created yet or that there is something terribly wrong with the installation. In either event, all the tables are recreated.